:arrow_up: Upgrade vuepress-theme-default to 1.x
