+++ docs/.vuepress/theme/noopModule.js
@@ -1,0 +1,1 @@
+export default {}
