+++ docs/.vuepress/theme/styles/mobile.styl
