+++ docs/.vuepress/theme/styles/palette.styl
@@ -1,0 +1,20 @@
+// colors
+$accentColor = #49B1F5
+$textColor = #2c3e50
+$borderColor = #eaecef
+$codeBgColor = #282C34
+$arrowBgColor = #ccc
+
+// layout
+$navbarHeight = 3.6rem
+$sidebarWidth = 20rem
+$contentWidth = 740px
+
+// responsive breakpoints
+$MQNarrow = 959px
+$MQMobile = 719px
+$MQMobileNarrow = 419px
+
+// code
+$lineNumbersWrapperWidth = 3.5rem
+$codeLang = js ts html md vue css sass scss less stylus go java c sh yaml py
