+++ docs/.vuepress/theme/styles/util/functions.styl
@@ -1,0 +1,2 @@
+px2rem($px, $fontSize = 37.5)
+  unit(($px / $fontSize), 'rem')
