+++ docs/.vuepress/theme/styles/util/mixnis.styl
@@ -1,0 +1,4 @@
+center()
+  display flex
+  justify-content center
+  align-items center
